feat: Update version to 1.1.1 and enhance Bitcoin mode functionality

+++ lib/features/dashboard/presentation/history_tab.dart
@@ -255,7 +255,7 @@ class HistoryTab extends ConsumerWidget {
 
   String _formatPrice(PurchaseEntry entry, AppSettings settings) {
     if (settings.isBitcoinMode && entry.priceSats != null) {
-      return SatsConverter.formatSats(entry.priceSats!.toInt());
+      return SatsConverter.formatSats(entry.priceSats!);
     }
     final currencyFormat = NumberFormat.simpleCurrency(name: settings.currency);
     return currencyFormat.format(entry.price);
